\name{get.master}
\alias{get.master}
\title{
  Return the master node from a cluster object.
}
\description{
  Return the master node info from a cluster.
}
\usage{
get.master(cluster)
}
\arguments{
  \item{cluster}{a cluster object}
}
\value{
  a list containing the details of the master node.
}
\references{
  http://docs.amazonwebservices.com/AWSEC2/latest/CommandLineReference
}
\author{
  Whit Armstrong
}
\keyword{utilities}
